+++ common/src/main/java/net/mehvahdjukaar/polytone/utils/CompoundReloader.java
@@ -5,6 +5,7 @@ import com.google.gson.JsonElement;
 import com.mojang.serialization.JsonOps;
 import net.mehvahdjukaar.polytone.Polytone;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.resources.RegistryOps;
 import net.minecraft.server.packs.resources.ResourceManager;
@@ -47,7 +48,7 @@ public class CompoundReloader extends SimplePreparableReloadListener<List<Object
         if (level != null) {
             try {
                 applyWithLevel(level.registryAccess(), false);
-            }catch (Exception e){
+            } catch (Exception e) {
                 throw new RuntimeException(e);
             }
         }
@@ -99,6 +100,10 @@ public class CompoundReloader extends SimplePreparableReloadListener<List<Object
                     player.containerMenu = player.inventoryMenu;
                 }
             }
+        }
+        Level level = Minecraft.getInstance().level;
+        if (level instanceof ClientLevel cl) {
+            cl.clearTintCaches();
         }
     }